requirements.txt
anyio==3.4.0
attrs==21.2.0
beautifulsoup4==4.10.0
blessed==1.19.0
botocore==1.23.26
bs4==0.0.1
certifi==2021.10.8
chalice==1.26.2
charset-normalizer==2.0.9
click==8.0.3
h11==0.12.0
httpcore==0.14.3
httpx==0.21.1
idna==3.3
inquirer==2.9.0
jmespath==0.10.0
mypy-extensions==0.4.3
python-dateutil==2.8.2
python-editor==1.0.4
python-slugify==5.0.2
PyYAML==6.0
readchar==3.0.4
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
text-unidecode==1.3
urllib3==1.26.7
wcwidth==0.2.5
